backend/requirements.txt
fastapi==0.115.0
mangum==0.17.0
pydantic==2.9.2
pydantic[email]==2.9.2
python-dotenv==1.0.1
requests==2.32.3
supabase==2.7.4
